Move CliArgs and main code to the mountpoint-s3 crate (#1401)
Complete the decoupling of the configuration code from the specific
CliArgs and initialization logic in the `mount-s3` binary. The latter
are now in the `mountpoint-s3` crate, while configuring Mountpoint is
now part of the `mountpoint-s3-fs` API.

### Does this change impact existing behavior?

No.

### Does this change need a changelog entry? Does it require a version
change?

Changelog entry for the `mountpoint-s3-fs` crate.

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------
