+++ mountpoint-s3-fs/src/lib.rs
@@ -1,7 +1,6 @@
 mod async_util;
 pub mod autoconfigure;
 mod checksums;
-pub mod cli;
 mod config;
 pub mod data_cache;
 pub mod fs;
