+++ mountpoint-s3/src/bin/mock-mount-s3.rs
@@ -13,26 +13,22 @@
 use std::sync::Arc;
 
 use anyhow::anyhow;
+use clap::Parser;
 use futures::executor::ThreadPool;
 
+use mountpoint_s3::CliArgs;
 use mountpoint_s3_client::mock_client::throughput_client::ThroughputMockClient;
 use mountpoint_s3_client::mock_client::{MockClientConfig, MockObject};
 use mountpoint_s3_client::types::ETag;
-use mountpoint_s3_fs::cli::{CliArgs, ContextParams};
 use mountpoint_s3_fs::s3::S3Personality;
+use mountpoint_s3_fs::Runtime;
 
 fn main() -> anyhow::Result<()> {
-    let cli_args = mountpoint_s3::try_parse_cli_args()?;
-    let context = ContextParams {
-        full_version: mountpoint_s3::build_info::FULL_VERSION.to_string(),
-    };
-    mountpoint_s3_fs::cli::main(create_mock_client, cli_args, context)
+    let cli_args = CliArgs::parse();
+    mountpoint_s3::run(create_mock_client, cli_args)
 }
 
-fn create_mock_client(
-    args: &CliArgs,
-    _context_params: &ContextParams,
-) -> anyhow::Result<(Arc<ThroughputMockClient>, ThreadPool, S3Personality)> {
+fn create_mock_client(args: &CliArgs) -> anyhow::Result<(Arc<ThroughputMockClient>, Runtime, S3Personality)> {
     // An extra little safety thing to make sure we can distinguish the real mount-s3 binary and
     // this one. Buckets starting with "sthree-" are always invalid against real S3:
     // https://docs.aws.amazon.com/AmazonS3/latest/userguide/bucketnamingrules.html
@@ -59,7 +55,7 @@ fn create_mock_client(
     };
     let client = ThroughputMockClient::new(config, max_throughput_gbps as f64);
 
-    let runtime = ThreadPool::builder().name_prefix("runtime").create()?;
+    let runtime = Runtime::new(ThreadPool::builder().name_prefix("runtime").create()?);
 
     let s3_personality = if let Some(bucket_type) = &args.bucket_type {
         bucket_type.to_personality()
