+++ mountpoint-s3/src/cli.rs
@@ -1,34 +1,25 @@
-use std::env;
-use std::ffi::OsString;
 use std::fmt::Debug;
-use std::fs::{self, File};
-use std::io::{Read, Write};
-use std::os::fd::AsRawFd;
 use std::path::PathBuf;
 use std::time::Duration;
 
 use anyhow::{anyhow, Context as _};
 use clap::{value_parser, ArgGroup, Parser, ValueEnum};
-use futures::task::Spawn;
-use mountpoint_s3_client::config::{AddressingStyle, EventLoopGroup, S3ClientAuthConfig, AWSCRT_LOG_TARGET};
+use mountpoint_s3_client::config::{AddressingStyle, S3ClientAuthConfig, AWSCRT_LOG_TARGET};
 use mountpoint_s3_client::instance_info::InstanceInfo;
 use mountpoint_s3_client::user_agent::UserAgent;
-use mountpoint_s3_client::{ObjectClient, S3CrtClient};
-use nix::sys::signal::Signal;
-use nix::unistd::ForkResult;
+use mountpoint_s3_fs::data_cache::{CacheLimit, DataCacheConfig, DiskDataCacheConfig, ExpressDataCacheConfig};
+use mountpoint_s3_fs::fs::{CacheConfig, ServerSideEncryption, TimeToLive};
+use mountpoint_s3_fs::fuse::config::{FuseOptions, FuseSessionConfig, MountPoint};
+use mountpoint_s3_fs::logging::{prepare_log_file_name, LoggingConfig};
+use mountpoint_s3_fs::mem_limiter::MINIMUM_MEM_LIMIT;
+use mountpoint_s3_fs::prefix::Prefix;
+use mountpoint_s3_fs::s3::config::{ClientConfig, PartConfig, S3Path};
+use mountpoint_s3_fs::s3::S3Personality;
+use mountpoint_s3_fs::{autoconfigure, metrics, S3FilesystemConfig};
 use regex::Regex;
 use sysinfo::{RefreshKind, System};
 
-use crate::data_cache::{CacheLimit, DataCacheConfig, DiskDataCacheConfig, ExpressDataCacheConfig, ManagedCacheDir};
-use crate::fs::{CacheConfig, ServerSideEncryption, TimeToLive};
-use crate::fuse::config::{FuseOptions, FuseSessionConfig, MountPoint};
-use crate::fuse::session::FuseSession;
-use crate::logging::{init_logging, prepare_log_file_name, LoggingConfig};
-use crate::mem_limiter::MINIMUM_MEM_LIMIT;
-use crate::prefix::Prefix;
-use crate::s3::config::{ClientConfig, PartConfig, S3Path};
-use crate::s3::S3Personality;
-use crate::{autoconfigure, metrics, MountpointConfig, Runtime, S3FilesystemConfig};
+use crate::build_info;
 
 const CLIENT_OPTIONS_HEADER: &str = "Client options";
 const MOUNT_OPTIONS_HEADER: &str = "Mount options";
@@ -38,17 +29,29 @@ const LOGGING_OPTIONS_HEADER: &str = "Logging options";
 const CACHING_OPTIONS_HEADER: &str = "Caching options";
 const ADVANCED_OPTIONS_HEADER: &str = "Advanced options";
 
-#[derive(Parser, Debug)]
-pub struct ContextParams {
-    pub full_version: String,
-}
+const FSTAB_DOCS: &str = "
+Alternative fstab style:
+  mount-s3 <BUCKET> <DIRECTORY> -o <OPTIONS>
+
+Arguments:
+  <BUCKET_NAME>
+          Name of bucket to mount, with s3:// URIs supported
+  <DIRECTORY>
+          Location to mount bucket at
+  <OPTIONS>
+          fstab style options. Comma separated list of CLI options, with backslash escapes for commas, backslashes, and double quotes.
+          Use of `--` to prefix arguments is not allowed.";
 
 #[derive(Parser, Debug)]
 #[clap(
+    name = "mount-s3",
+    about = "Mountpoint for Amazon S3",
+    version = build_info::FULL_VERSION,
     group(
         ArgGroup::new("cache_group")
             .multiple(true),
     ),
+    after_help = if cfg!(feature = "fstab") {FSTAB_DOCS} else {""},
 )]
 pub struct CliArgs {
     #[clap(help = "Name of bucket to mount", value_parser = parse_bucket_name)]
@@ -393,7 +396,7 @@ pub enum BucketType {
 }
 
 impl BucketType {
-    pub fn to_personality(&self) -> S3Personality {
+    pub fn to_personality(self) -> S3Personality {
         match self {
             Self::GeneralPurpose => S3Personality::Standard,
             Self::Directory => S3Personality::ExpressOneZone,
@@ -446,7 +449,7 @@ impl CliArgs {
         self.prefix.as_ref().cloned().unwrap_or_default()
     }
 
-    fn s3_path(&self) -> S3Path {
+    pub fn s3_path(&self) -> S3Path {
         S3Path::new(self.bucket_name.to_owned(), self.prefix())
     }
 
@@ -482,7 +485,7 @@ impl CliArgs {
         }
     }
 
-    fn filesystem_config(&self, sse: ServerSideEncryption, s3_personality: S3Personality) -> S3FilesystemConfig {
+    pub fn filesystem_config(&self, sse: ServerSideEncryption, s3_personality: S3Personality) -> S3FilesystemConfig {
         let mut filesystem_config = S3FilesystemConfig::default();
         if let Some(uid) = self.uid {
             filesystem_config.uid = uid;
@@ -580,7 +583,7 @@ impl CliArgs {
         Some(cache_config)
     }
 
-    fn data_cache_config(&self, sse: ServerSideEncryption) -> DataCacheConfig {
+    pub fn data_cache_config(&self, sse: ServerSideEncryption) -> DataCacheConfig {
         let disk_cache_config = self.disk_data_cache_config();
         let express_cache_config = self.express_data_cache_config(sse);
         match (&disk_cache_config, &express_cache_config) {
@@ -606,7 +609,7 @@ impl CliArgs {
     /// The server-side encryption configuration.
     ///
     /// Disallow specifying `--sse-kms-key-id` when `--sse=AES256` as this is not allowed by the S3 API.
-    fn server_side_encryption(&self) -> anyhow::Result<ServerSideEncryption> {
+    pub fn server_side_encryption(&self) -> anyhow::Result<ServerSideEncryption> {
         if self.sse_kms_key_id.is_some() && self.sse.as_deref() == Some("AES256") {
             return Err(anyhow!("--sse-kms-key-id can not be used with --sse AES256"));
         }
@@ -617,7 +620,7 @@ impl CliArgs {
     ///
     /// This includes random string generation which can change with each invocation,
     /// so once created the [LoggingConfig] should be cloned if another owned copy is required.
-    fn make_logging_config(&self) -> LoggingConfig {
+    pub fn make_logging_config(&self) -> LoggingConfig {
         let default_filter = if self.no_log {
             String::from("off")
         } else {
@@ -644,7 +647,7 @@ impl CliArgs {
     }
 
     /// Human-readable description of the bucket being mounted
-    fn bucket_description(&self) -> String {
+    pub fn bucket_description(&self) -> String {
         if let Some(prefix) = self.prefix.as_ref() {
             format!("prefix {} of bucket {}", prefix, self.bucket_name)
         } else {
@@ -652,7 +655,7 @@ impl CliArgs {
         }
     }
 
-    fn fuse_session_config(&self) -> anyhow::Result<FuseSessionConfig> {
+    pub fn fuse_session_config(&self) -> anyhow::Result<FuseSessionConfig> {
         let mount_point = MountPoint::new(&self.mount_point).context("Failed to create mount point")?;
         let fuse_options = FuseOptions {
             read_only: self.read_only,
@@ -713,7 +716,7 @@ impl CliArgs {
         throughput_target_gbps
     }
 
-    fn personality(&self) -> Option<S3Personality> {
+    pub fn personality(&self) -> Option<S3Personality> {
         self.bucket_type.map(|bucket_type| bucket_type.to_personality())
     }
 
@@ -734,7 +737,7 @@ impl CliArgs {
         }
     }
 
-    fn client_config(&self, version: &str) -> ClientConfig {
+    pub fn client_config(&self, version: &str) -> ClientConfig {
         let instance_info = InstanceInfo::new();
         let user_agent = self.user_agent(&instance_info, version);
         let throughput_target_gbps = self.throughput_target_gbps(&instance_info);
@@ -755,255 +758,6 @@ impl CliArgs {
             user_agent,
         }
     }
-}
-
-pub fn main<ClientBuilder, Client, Runtime>(
-    client_builder: ClientBuilder,
-    args: CliArgs,
-    context_params: ContextParams,
-) -> anyhow::Result<()>
-where
-    ClientBuilder: FnOnce(&CliArgs, &ContextParams) -> anyhow::Result<(Client, Runtime, S3Personality)>,
-    Client: ObjectClient + Clone + Send + Sync + 'static,
-    Runtime: Spawn + Clone + Send + Sync + 'static,
-{
-    let successful_mount_msg = format!(
-        "{} is mounted at {}",
-        args.bucket_description(),
-        args.mount_point.display()
-    );
-
-    if args.foreground {
-        let _logging = init_logging(args.make_logging_config()).context("failed to initialize logging")?;
-
-        let _metrics = metrics::install();
-
-        create_pid_file()?;
-
-        // mount file system as a foreground process
-        let session = mount(args, client_builder, context_params)?;
-
-        println!("{successful_mount_msg}");
-
-        session.join().context("failed to join session")?;
-    } else {
-        // mount file system as a background process
-
-        // create a pipe for interprocess communication.
-        // child process will report its status via this pipe.
-        let (read_fd, write_fd) = nix::unistd::pipe().context("Failed to create a pipe")?;
-
-        // Prepare logging configuration up front, so the values are shared between the two processes.
-        let logging_config = args.make_logging_config();
-
-        // Don't share args across the fork. It should just be plain data, so probably fine to be
-        // copy-on-write, but just in case we ever add something more fancy to the struct.
-        drop(args);
-
-        // SAFETY: Child process has full ownership of its resources.
-        // There is no shared data between parent and child processes other than logging configuration.
-        let pid = unsafe { nix::unistd::fork() };
-        match pid.expect("Failed to fork mount process") {
-            ForkResult::Child => {
-                let args = CliArgs::parse();
-                let _logging = init_logging(logging_config).context("failed to initialize logging")?;
-
-                let _metrics = metrics::install();
-
-                create_pid_file()?;
-
-                let session = mount(args, client_builder, context_params);
-
-                // close unused file descriptor, we only write from this end.
-                drop(read_fd);
-
-                let mut pipe_file = File::from(write_fd);
-
-                let status_success = [b'0'];
-                let status_failure = [b'1'];
-
-                match session {
-                    Ok(session) => {
-                        tracing::trace!("FUSE session created OK, sending message back to parent process");
-                        pipe_file
-                            .write(&status_success)
-                            .context("Failed to write data to the pipe")?;
-                        drop(pipe_file);
-                        tracing::trace!("message sent back to parent process");
-
-                        // Logging is set up and the mount succeeded, so we can hang up
-                        // stdin/out/err now to cleanly daemonize ourselves
-                        nix::unistd::close(std::io::stdin().as_raw_fd()).context("couldn't close stdin")?;
-                        nix::unistd::close(std::io::stdout().as_raw_fd()).context("couldn't close stdout")?;
-                        nix::unistd::close(std::io::stderr().as_raw_fd()).context("couldn't close stderr")?;
-
-                        session.join().context("failed to join session")?;
-                    }
-                    Err(e) => {
-                        tracing::trace!("FUSE session creation failed, sending message back to parent process");
-                        pipe_file
-                            .write(&status_failure)
-                            .context("Failed to write data to the pipe")?;
-                        tracing::trace!("message sent back to parent process");
-                        return Err(anyhow!(e));
-                    }
-                }
-            }
-            ForkResult::Parent { child } => {
-                let _logging = init_logging(logging_config).context("failed to initialize logging")?;
-
-                // close unused file descriptor, we only read from this end.
-                drop(write_fd);
-
-                let mut pipe_file = File::from(read_fd);
-
-                let (sender, receiver) = std::sync::mpsc::channel();
-
-                // create a thread that read from the pipe so that we can enforce a time out.
-                std::thread::spawn(move || {
-                    let mut buf = [0];
-                    match pipe_file
-                        .read_exact(&mut buf)
-                        .context("Failed to read data from the pipe")
-                    {
-                        Ok(_) => {
-                            let status = buf[0] as char;
-                            sender.send(status).unwrap();
-                        }
-                        Err(_) => sender.send('1').unwrap(),
-                    }
-                });
-
-                let timeout = Duration::from_secs(30);
-                tracing::debug!(
-                    "waiting up to {} seconds for child process to be ready",
-                    timeout.as_secs(),
-                );
-                let status = receiver.recv_timeout(timeout);
-                match status {
-                    Ok('0') => {
-                        println!("{successful_mount_msg}");
-                        tracing::debug!("success status flag received from child process")
-                    }
-                    Ok(_) => {
-                        nix::sys::wait::waitpid(child, None).context("Failed to wait for child process to exit")?;
-                        return Err(anyhow!("Failed to create mount process"));
-                    }
-                    Err(_timeout_err) => {
-                        tracing::error!(
-                            "timeout after {} seconds waiting for message from child process",
-                            timeout.as_secs(),
-                        );
-                        // kill child process before returning error.
-                        if let Err(e) = nix::sys::signal::kill(child, Signal::SIGTERM) {
-                            tracing::error!("Unable to kill hanging child process with SIGTERM: {:?}", e);
-                        }
-                        return Err(anyhow!(
-                            "Timeout after {} seconds while waiting for mount process to be ready",
-                            timeout.as_secs()
-                        ));
-                    }
-                };
-            }
-        }
-    }
-
-    Ok(())
-}
-
-/// Create a real S3 client
-pub fn create_s3_client(
-    args: &CliArgs,
-    context_params: &ContextParams,
-) -> anyhow::Result<(S3CrtClient, EventLoopGroup, S3Personality)> {
-    // We keep this logic here until we decouple config layer from FS implementation
-    // Once we do that, we can move this logic into a hosting app as it will know the full context
-    // and remove the contextParams struct
-    let version = &context_params.full_version;
-    let client_config = args.client_config(version);
-
-    let s3_path = args.s3_path();
-    let client = client_config
-        .create_client(Some(&s3_path))
-        .context("Failed to create S3 client")?;
-
-    let runtime = client.event_loop_group();
-    let s3_personality = args
-        .personality()
-        .unwrap_or_else(|| S3Personality::infer_from_bucket(&s3_path.bucket_name, &client.endpoint_config()));
-
-    Ok((client, runtime, s3_personality))
-}
-
-fn mount<ClientBuilder, Client, ClientRuntime>(
-    args: CliArgs,
-    client_builder: ClientBuilder,
-    context_params: ContextParams,
-) -> anyhow::Result<FuseSession>
-where
-    ClientBuilder: FnOnce(&CliArgs, &ContextParams) -> anyhow::Result<(Client, ClientRuntime, S3Personality)>,
-    Client: ObjectClient + Clone + Send + Sync + 'static,
-    ClientRuntime: Spawn + Clone + Send + Sync + 'static,
-{
-    tracing::info!("mount-s3 {}", context_params.full_version);
-    tracing::debug!("{:?}", args);
-
-    let fuse_session_config = args.fuse_session_config()?;
-    let sse = args.server_side_encryption()?;
-
-    let (client, runtime, s3_personality) = client_builder(&args, &context_params)?;
-    let runtime = Runtime::new(runtime);
-
-    let bucket_description = args.bucket_description();
-    tracing::debug!("using S3 personality {s3_personality:?} for {bucket_description}");
-
-    let s3_path = args.s3_path();
-    let filesystem_config = args.filesystem_config(sse.clone(), s3_personality);
-    let mut data_cache_config = args.data_cache_config(sse);
-
-    let managed_cache_dir = setup_disk_cache_directory(&mut data_cache_config)?;
-
-    tracing::debug!(?fuse_session_config, "creating fuse session");
-    let mount_point_path = format!("{}", fuse_session_config.mount_point());
-
-    let mut fuse_session = MountpointConfig::new(fuse_session_config, filesystem_config, data_cache_config)
-        .create_fuse_session(s3_path, client, runtime)?;
-    tracing::info!("successfully mounted {} at {}", bucket_description, mount_point_path);
-
-    if let Some(managed_cache_dir) = managed_cache_dir {
-        fuse_session.run_on_close(Box::new(move || {
-            drop(managed_cache_dir);
-        }));
-    }
-    Ok(fuse_session)
-}
-
-fn setup_disk_cache_directory(cache_config: &mut DataCacheConfig) -> anyhow::Result<Option<ManagedCacheDir>> {
-    let Some(disk_cache_config) = &mut cache_config.disk_cache_config else {
-        return Ok(None);
-    };
-    let cache_key = env_unstable_cache_key();
-    let managed_cache_dir =
-        ManagedCacheDir::new_from_parent_with_cache_key(&disk_cache_config.cache_directory, cache_key.as_deref())
-            .context("failed to create cache directory")?;
-    disk_cache_config.cache_directory = managed_cache_dir.as_path_buf();
-    Ok(Some(managed_cache_dir))
-}
-
-/// Creates PID file at location specified by env var, writing the PID of the Mountpoint process.
-///
-/// The written PID may not match the PID visible in your namespace.
-/// This can happen, for example, when using it from the host when Mountpoint runs in a container.
-///
-/// PID file configuration is available for attaching debug tooling to Mountpoint, and may be removed in the future.
-fn create_pid_file() -> anyhow::Result<()> {
-    const ENV_PID_FILENAME: &str = "UNSTABLE_MOUNTPOINT_PID_FILE";
-    if let Some(val) = std::env::var_os(ENV_PID_FILENAME) {
-        let pid = std::process::id();
-        fs::write(&val, pid.to_string()).context("failed to write PID to file")?;
-        tracing::trace!("PID ({pid}) written to file {val:?}");
-    }
-    Ok(())
 }
 
 fn parse_perm_bits(perm_bit_str: &str) -> Result<u16, anyhow::Error> {
@@ -1048,10 +802,6 @@ fn parse_kms_key_arn(kms_key_arn: &str) -> anyhow::Result<String> {
             "KMS Key ARN is only accepted as a key identifier, Key Alias ARN is not accepted"
         ))
     }
-}
-
-fn env_unstable_cache_key() -> Option<OsString> {
-    env::var_os("UNSTABLE_MOUNTPOINT_CACHE_KEY")
 }
 
 #[cfg(test)]
