+++ mountpoint-s3/src/main.rs
@@ -1,16 +1,19 @@
-use mountpoint_s3::build_info;
-use mountpoint_s3::try_parse_cli_args;
+use std::env;
+
+use clap::Parser;
+
+use mountpoint_s3::{create_s3_client, CliArgs, FsTabCliArgs};
 
 fn main() -> anyhow::Result<()> {
-    match try_parse_cli_args() {
-        Ok(cli_args) => {
-            let context = mountpoint_s3_fs::cli::ContextParams {
-                full_version: build_info::FULL_VERSION.to_string(),
-            };
-            mountpoint_s3_fs::cli::main(mountpoint_s3_fs::cli::create_s3_client, cli_args, context)
-        }
-        Err(err) => {
-            err.exit();
-        }
-    }
+    let is_fstab = env::args_os().len() == 5 && env::args_os().nth(3).as_deref() == Some("-o".as_ref());
+
+    let cli_args = if is_fstab && cfg!(feature = "fstab") {
+        println!("Using 'fstab' style options as detected use of `-o` argument.");
+        FsTabCliArgs::try_parse().and_then(|args| args.try_into())
+    } else {
+        CliArgs::try_parse()
+    };
+
+    let cli_args = cli_args.unwrap_or_else(|err| err.exit());
+    mountpoint_s3::run(create_s3_client, cli_args)
 }
